severas mudanças no dominio e recursos

+++ src/main/java/com/pequenosgenios/pg/repositories/ClassRepository.java
@@ -1,9 +1,11 @@
 package com.pequenosgenios.pg.repositories;
 
 import com.pequenosgenios.pg.domain.Class;
+import com.pequenosgenios.pg.domain.Student;
+import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassRepository extends JpaRepository<Class, Integer> {
+public interface ClassRepository extends JpaRepository<Class, Long> {
 }
